fix compilation with glibc 2.28
In glibc 2.28 function fadd, fsub etc. are added to math.h [0].  Miracl
defines functions with the same names.  Hence, inclusion of both, math.h
and Miracl, in the same translation unit results in an error.

To address this issue, the use of Miracl is moved to ecc-pk-crypto.cpp
such that inclusion of the header does not include Miracl anymore.
Therefore, math.h can now be included together with ecc-pk-crypto.h in
other source files (as long as fadd etc. from math.h are not used
anywhere).

This change was already committed in upstream ENCRYPTO_utils [1].

[0]: https://savannah.gnu.org/forum/forum.php?forum_id=9205
[1]: https://github.com/encryptogroup/ENCRYPTO_utils/commit/f0b707c48294e84d35a6ef31d1f31b21a90a0c1a